warning reduction and elimination mostly associated with long types
git-svn-id: http://svn.gekkoware.net/gas/trunk@73 e362cfe6-bd3b-0410-998e-f663b8f83dfe

+++ src/gas/context.h
@@ -34,7 +34,7 @@ typedef GASnum (*GAS_FILE_WRITE_CALLBACK) (void *handle, void *buffer,
  * @brief provides the ability to seek forward (SEEK_CUR).
  */
 typedef GASnum (*GAS_FILE_SEEK_CALLBACK)  (void *handle,
-                                           unsigned int pos,
+                                           unsigned long pos,
                                            int whence,
                                            void *userdata);
